requirements.txt
discord.py==1.4.1
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1